added universal component CorgiActions, updated indexes, removed code duplicates
