--- src/modules/common/corgi/actions/ActionsDropdown/index.js
@@ -1,3 +1,0 @@
-import ActionsDropdown from './ActionsDropdown';
-
-export default ActionsDropdown;
