+++ src/modules/common/corgi/actions/index.js
@@ -1,7 +1,7 @@
-import ActionsDropdown from './ActionsDropdown';
+import CorgiActions from './CorgiActions';
 import Confirmation from './Confirmation';
 import Share from './Share';
 import Transfer from './Transfer';
 import { DeletePopup, GiftPopup, SharePopup, TradePopup } from './popups';
 
-export { ActionsDropdown, Confirmation, Share, Transfer, DeletePopup, GiftPopup, SharePopup, TradePopup };
+export { CorgiActions, Confirmation, Share, Transfer, DeletePopup, GiftPopup, SharePopup, TradePopup };
