+++ src/modules/common/corgi/index.js
@@ -2,16 +2,7 @@ import { CorgiAnimOne, CorgiAnimTwo, CorgiAnimThree, CorgiAnimFour } from './Cor
 
 import { CorgiCommon, CorgiUncommon, CorgiRare, CorgiVeryRare, SwitchCorgiPhoto } from './CorgiPhotos';
 
-import {
-  ActionsDropdown,
-  Confirmation,
-  Share,
-  Transfer,
-  DeletePopup,
-  GiftPopup,
-  SharePopup,
-  TradePopup,
-} from './actions';
+import { CorgiActions, Confirmation, Share, Transfer, DeletePopup, GiftPopup, SharePopup, TradePopup } from './actions';
 
 import CorgiCard from './CorgiCard';
 import CorgiLink from './CorgiLink';
@@ -46,7 +37,7 @@ export {
   Owner,
   RarityString,
   Quote,
-  ActionsDropdown,
+  CorgiActions,
   Confirmation,
   Share,
   Transfer,
