--- src/modules/corgi/components/index.js
@@ -1,3 +1,0 @@
-import CorgiActions from './CorgiActions';
-
-export { CorgiActions };
